pyproject.toml
[project]
name = "anonimous"
version = "0.1.0"
description = "O Anonymizer é uma ferramenta que oferece anonimização digital instantânea para imagens, eliminando metadados e rastros ocultos, garantindo privacidade total e segurança."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "pillow>=12.2.0",
]
